# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Creator AI Backend Engine",
    "description": "\n# Creator AI \u2014 Autonomous Creator Backend API Engine\n\nHigh-performance backend for the React Native Creator AI mobile application, delivering real-time multimodal creator intelligence:\n\n- **`/profiling`**: Ingests real creator videos, shorts, transcripts, Substack newsletters, and social posts. Deeply extracts tone, video duration distribution, video types, thumbnail strategies, and authentic spoken vocal mannerisms (\"what user says often\"), generating `user.md` and `hook.md`.\n- **`/dashboard`**: Unified cross-platform metrics across YouTube, Substack, LinkedIn, and X/Twitter + in-app Creator AI pipeline analytics and React Native chart datasets.\n- **`/trends`**: Live Google Trends RSS feeds, domain/niche-specific trending topics, viral format templates, and actionable content angles.\n- **`/publish`**: Human-In-The-Loop approval gate with **Composio** automation for publishing to YouTube, LinkedIn, X/Twitter, and Substack.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/trends": {
      "get": {
        "tags": [
          "Live Trends & Hook Angles"
        ],
        "summary": "Get trendy videos/posts for domain across YouTube, Instagram, LinkedIn & X",
        "description": "Fetches live real-world trending content across all 4 major platforms (YouTube, Instagram, LinkedIn, X/Twitter)\nfor the specified domain, plus high-velocity trending keywords, viral format templates, and ready-to-record hook angles.",
        "operationId": "get_trends_trends_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Domain/niche (e.g. 'Tech & AI', 'Finance', 'Productivity', 'Fitness', 'Social Causes')",
              "default": "Tech & AI",
              "title": "Domain"
            },
            "description": "Domain/niche (e.g. 'Tech & AI', 'Finance', 'Productivity', 'Fitness', 'Social Causes')"
          },
          {
            "name": "creator_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional creator name to tailor niche hooks",
              "title": "Creator Name"
            },
            "description": "Optional creator name to tailor niche hooks"
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Geographic region code for Google Trends (e.g. 'US', 'GB', 'IN', 'GLOBAL')",
              "default": "US",
              "title": "Geo"
            },
            "description": "Geographic region code for Google Trends (e.g. 'US', 'GB', 'IN', 'GLOBAL')"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 3,
              "description": "Number of trending topics to return per platform",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of trending topics to return per platform"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trends/domain": {
      "get": {
        "tags": [
          "Live Trends & Hook Angles"
        ],
        "summary": "Search domain for trendy videos/posts across YouTube, Instagram, LinkedIn & X",
        "description": "Search any domain to uncover real-time YouTube trending videos, viral Instagram Reels,\nhigh-authority LinkedIn discussions, viral X/Twitter threads, and high-velocity trending keywords.",
        "operationId": "search_domain_trends_get_trends_domain_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain/niche (e.g. 'Fitness', 'Gaming', 'Crypto', 'AI Agents', 'Finance', 'Social Causes')",
              "title": "Domain"
            },
            "description": "Domain/niche (e.g. 'Fitness', 'Gaming', 'Crypto', 'AI Agents', 'Finance', 'Social Causes')"
          },
          {
            "name": "geo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Geographic region code ('US', 'IN', 'GLOBAL')",
              "default": "US",
              "title": "Geo"
            },
            "description": "Geographic region code ('US', 'IN', 'GLOBAL')"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 3,
              "description": "Number of results per category",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of results per category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/intelligence": {
      "post": {
        "tags": [
          "Platform Intelligence & Trends"
        ],
        "summary": "Run full cross-platform intelligence scan",
        "description": "Performs a comprehensive trend analysis across selected platforms (YouTube, Instagram, LinkedIn, X/Twitter).\n\n**What it does:**\n1. Scrapes live trending content per platform for the creator's niche\n2. Discovers domain, location, and global trends with links\n3. Generates platform-specific content strategies\n4. Produces AI-powered content recommendations\n5. Auto-generates `platform_{name}.md` files under `creators/{slug}/`\n\n**Platform Goals:**\n- Instagram \u2192 Increase Reach\n- YouTube \u2192 Increase Followers\n- LinkedIn \u2192 Increase Connections\n- X/Twitter \u2192 Increase Engagement",
        "operationId": "run_intelligence_scan_intelligence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/intelligence/quick": {
      "get": {
        "tags": [
          "Platform Intelligence & Trends"
        ],
        "summary": "Quick intelligence scan via query params",
        "description": "Lightweight GET endpoint for quick trend scanning without a JSON body.",
        "operationId": "quick_intelligence_scan_intelligence_quick_get",
        "parameters": [
          {
            "name": "creator_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Creator's display name",
              "title": "Creator Name"
            },
            "description": "Creator's display name"
          },
          {
            "name": "niche",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Creator's niche/domain",
              "default": "Tech & AI",
              "title": "Niche"
            },
            "description": "Creator's niche/domain"
          },
          {
            "name": "causes_or_topics",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated topics or causes of creator's work (e.g. 'Social Causes, Public Policy')",
              "title": "Causes Or Topics"
            },
            "description": "Comma-separated topics or causes of creator's work (e.g. 'Social Causes, Public Policy')"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Creator's native spoken language code (e.g. 'hi', 'en', 'es')",
              "title": "Language"
            },
            "description": "Creator's native spoken language code (e.g. 'hi', 'en', 'es')"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Country/region code",
              "default": "US",
              "title": "Location"
            },
            "description": "Country/region code"
          },
          {
            "name": "platforms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated platform names (youtube, instagram, linkedin, x_twitter)",
              "default": "youtube,instagram,linkedin,x_twitter",
              "title": "Platforms"
            },
            "description": "Comma-separated platform names (youtube, instagram, linkedin, x_twitter)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/intelligence/identify-domain": {
      "post": {
        "tags": [
          "Platform Intelligence & Trends"
        ],
        "summary": "Identify creator domain, niche, and content pillars",
        "description": "Dynamically identifies any creator's domain, core verticals, target audience psychographics, competitive moat, and domain monologues.",
        "operationId": "identify_creator_domain_intelligence_identify_domain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentifyDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorDomainProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/intelligence/compare": {
      "post": {
        "tags": [
          "Platform Intelligence & Trends"
        ],
        "summary": "Compare with any creator across any domain, extracting all 7 dossiers",
        "description": "Takes input of any creator (no matter what domain), extracts all platform.md files, user.md, hook.md, and creator_comparison.md, and benchmarks against our base creator profile.",
        "operationId": "compare_any_creator_intelligence_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareAnyCreatorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompareAnyCreatorRequest": {
        "properties": {
          "creator_name": {
            "type": "string",
            "title": "Creator Name",
            "description": "Target creator name (e.g. 'Lex Fridman', 'Veritasium', 'MrBeast')"
          },
          "base_creator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Creator",
            "description": "Base creator to benchmark against ('me')",
            "default": "Dhruv Rathee"
          },
          "niche_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche Hint",
            "description": "Optional hint for target creator's domain"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Location country code",
            "default": "US"
          }
        },
        "type": "object",
        "required": [
          "creator_name"
        ],
        "title": "CompareAnyCreatorRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentifyDomainRequest": {
        "properties": {
          "creator_name": {
            "type": "string",
            "title": "Creator Name",
            "description": "Creator's display name or handle"
          },
          "niche_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche Hint",
            "description": "Optional hint about niche if known"
          },
          "sample_titles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Titles",
            "description": "Recent video/post titles to analyze"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio",
            "description": "Channel or profile bio text"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Preferred language for monologues and hooks ('hi', 'en', 'es')",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "creator_name"
        ],
        "title": "IdentifyDomainRequest"
      },
      "IntelligenceRequest": {
        "properties": {
          "creator_name": {
            "type": "string",
            "title": "Creator Name",
            "description": "Creator's display name or handle"
          },
          "niche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche",
            "description": "Creator's primary niche/domain (e.g. 'Tech & Gadgets', 'Finance & Investing', 'Social Causes', 'Productivity', or 'auto' for automatic detection)",
            "default": "auto"
          },
          "causes_or_topics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causes Or Topics",
            "description": "Creator's specific causes or topics of work (e.g. ['Environmental Crisis', 'Civic Awareness', 'Public Policy']) to find trends related to their work rather than personal news"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Creator's native spoken language code (e.g. 'hi' for Hindi/Hinglish, 'en' for English, 'es' for Spanish). Auto-detected if omitted."
          },
          "location": {
            "type": "string",
            "title": "Location",
            "description": "Country/region code for location-based trends (e.g. 'US', 'IN', 'GB')",
            "default": "US"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformChoice"
            },
            "type": "array",
            "title": "Platforms",
            "description": "Which platforms to investigate. Only the specified platforms will be queried.",
            "default": [
              "youtube",
              "instagram",
              "linkedin",
              "x_twitter"
            ]
          },
          "platform_handles": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Handles",
            "description": "Optional handles per platform, e.g. {'youtube': '@dhruvrathee', 'instagram': 'dhruvrathee', 'linkedin': 'dhruvrathee', 'x_twitter': 'dhruv_rathee'}"
          },
          "goals": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals",
            "description": "Per-platform goals (e.g. {'youtube': 'increase_subscribers', 'instagram': 'more_views_and_followers'})"
          },
          "generate_platform_md": {
            "type": "boolean",
            "title": "Generate Platform Md",
            "description": "Whether to auto-generate platform.md files",
            "default": true
          },
          "generate_user_hook_md": {
            "type": "boolean",
            "title": "Generate User Hook Md",
            "description": "Whether to generate/update user.md and hook.md in the creator's native language based on the investigated apps",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "creator_name"
        ],
        "title": "IntelligenceRequest"
      },
...
      "PlatformChoice": {
        "type": "string",
        "enum": [
          "youtube",
          "instagram",
          "linkedin",
          "x_twitter"
        ],
        "title": "PlatformChoice"
      },
...
      "TrendsResponse": {
        "properties": {
          "creator_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "geo": {
            "type": "string",
            "title": "Geo"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          },
          "youtube_trending": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Youtube Trending",
            "description": "Trending videos and Shorts in this domain"
          },
          "instagram_trending": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Instagram Trending",
            "description": "Viral Instagram Reels and posts in this domain"
          },
          "linkedin_trending": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Linkedin Trending",
            "description": "High-engagement LinkedIn discussions and carousels in this domain"
          },
          "x_twitter_trending": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "X Twitter Trending",
            "description": "Viral X / Twitter tweets and threads in this domain"
          },
          "trending_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trending Keywords",
            "description": "High-velocity search and discussion keywords in this domain"
          },
          "velocity_topics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Velocity Topics",
            "description": "Surging search queries and industry velocity topics"
          },
          "niche_trends": {
            "items": {
              "$ref": "#/components/schemas/TrendingItem"
            },
            "type": "array",
            "title": "Niche Trends"
          },
          "world_trends": {
            "items": {
              "$ref": "#/components/schemas/TrendingItem"
            },
            "type": "array",
            "title": "World Trends"
          },
          "viral_formats": {
            "items": {
              "$ref": "#/components/schemas/FormatTrend"
            },
            "type": "array",
            "title": "Viral Formats"
          },
          "content_opportunity_matrix": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Content Opportunity Matrix"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "geo",
          "fetched_at"
        ],
        "title": "TrendsResponse"
      },
...
    }
  }
}